Unionfs: branch-management bug-fix to lower mnt counts
When we revalidate opened files, the files may point to a different set of
files on the lower branches: this could happen especially if we insert a new
leftmost branch which happens to have duplicate file names of already-open
files in lower-priority branches.  In this case, update the mnt counts of
the old and new branches which hold the file accordingly.  Without this fix,
the mntcnt of one branch is going to be one too high; and another will be
one too low.

Signed-off-by: Erez Zadok <ezk@cs.sunysb.edu>